+++ src/services/api/fantlab/book.ts
@@ -80,17 +80,20 @@ function children(w) {
   }));
 }
 
-function editionIds(w) {
+function getEditions(w) {
   const RUSSIAN_EDITION = 10
-  return _.get(w, `editions_blocks.${RUSSIAN_EDITION}.list`, []).map(el => el.edition_id)
+  return _.get(w, `editions_blocks.${RUSSIAN_EDITION}.list`, [])
+}
+
+function editionIds(w) {
+  return getEditions(w).map(el => el.edition_id)
 }
 
 function translators(w) {
-  const editions = _.get(w, 'editions_blocks.10.list', [])
   const translators = _.get(w, 'editions_info.translators', [])
   const translatorNames = {}
 
-  _.forEach(editions, el => {
+  _.forEach(getEditions(w), el => {
     translatorNames[el.edition_id] = el.translators.split(',').map(id => _.get(translators.find(t => t.id === id), 'name', ''))
   })
