+++ src/services/api/fantlab/editions.ts
@@ -2,8 +2,6 @@ import _ from 'lodash';
 
 export const url = '/search-ids';
 
-export const cache = true;
-
 export function mapParams({ e }) {
   return {
     query: { e },
@@ -16,4 +14,16 @@ export const mapBody = {
 
 function editions(e) {
   return _.map(e, (el) => ({ ...el, thumbnail: el.image ? `https:${el.image}` : null, }))
+}
+
+export interface Edition {
+  copies: number
+  id: number
+  image: string
+  isbns: string[]
+  thumbnail: string
+  pages: number
+  published: number
+  url: string
+  year: number
 }
